requirements.txt
numpy
pandas
matplotlib
scipy
statsmodels
scikit-learn
stanza
networkx
praw
python-docx
pygame